refactor auth check and helpers
