+++ api/helper.js
@@ -1,25 +1,15 @@
 var createError = require('http-errors');
+var logger = require('../logger');
 var dataService = require('./dataService');
 
 module.exports = {
-  missingAuthHeader (req) {
-    return !(req.header('Authorization') || req.header('authorization'))
-  },
+  checkAuthHeader (req, res, next) {
+    // don't require auth for reads
+    if (/get/i.test(req.method)) next();
 
-  // look for user's post, hand to callback or return 404
-  postAction(req, res, callback) {
-    const userId = Number(req.params.userId);
-    const postId = Number(req.params.postId);
-    const post = dataService.getPost(userId, postId);
-    if (post) return callback(res, post);
-    else res.status(404).json(createError(404, `Post ${postId} by ${userId} does not exist!`));
-  },
-
-  // look for user, hand to callback or return 404
-  userAction(req, res, callback) {
-    const userId = Number(req.params.userId);
-    const user = dataService.getUser(userId);
-    if (user) return callback(res, user);
-    else res.status(404).json(createError(404, `User ${userId} does not exist!`));
+    if (!(req.header('Authorization') || req.header('authorization'))) {
+      logger.warn('Unauthorized access attempted!');
+      res.status(401).json(createError(401, 'Not Authorized'));
+    } else next();
   }
 }
